requirements.txt
python-dotenv
langchain
langchain-openai
langchain-chroma
langchain-community
numpy
scikit-learn
plotly
openai
gradio